update loading + pagination
